+++ vscode-ext/glyph-middleware.js
@@ -1230,7 +1230,7 @@ class GlyphCompressor {
 
   _createSourceMap() {
     return {
-      version: '1.30.1',
+      version: '1.31.0',
       level: this.level,
       provider: this.provider,
       profile: this.providerProfile,